+++ main.tf
@@ -63,13 +63,19 @@ resource "aws_instance" "example" {
 # ---------------------------------------------------------------------------------------------------------------------
 
 resource "aws_security_group" "instance" {
-  name = "terraform-example-instance"
+  name        = "terraform-example-instance"
+  description = "Allow HTTP access to the Terraform example web server"
 
   # Inbound HTTP from anywhere
   ingress {
+    description = "HTTP access to the example web server"
     from_port   = var.server_port
     to_port     = var.server_port
     protocol    = "tcp"
     cidr_blocks = var.allowed_cidr_blocks
+  }
+
+  tags = {
+    Name = "terraform-example-instance"
   }
 }